refactor: NotificationHelper 클래스를 제거하고 Extension으로 이동
Related to : #251
